+++ client/src/App.jsx
@@ -1,7 +1,7 @@
 import "./App.css";
 import "bootstrap/dist/css/bootstrap.min.css";
 import MyNavBar from "./components/Navbar";
-import ListJobs from "./components/ListStudents";
+import ListJobs from "./components/ListJobs";
 
 function App() {
   return (
